+++ ReleaseNotes.md
@@ -24,6 +24,7 @@ SOFTWARE.
 
 * Partial MISRA C2012 compliance
 * Moved from TravisCI to Github workflows
+* OVF_DIFF removal
 * arm_macros.h: BYTE_TO_PERC test >= 255 which doesn't raise warning when given BYTE parameter
 * arm_macros.h: re-enabled TEST_BITS_SET & TEST_BITS_CLR macros (TEST_BITS_VAL set to be alias of TEST_BITS_SET)
 * arm_keywords.h: \__NEVER_INLINE symbol added